Create view to fake ERA5 forecast analysis times
ERA5 has a different analysis time for each hour for instant parameters
with 0-length forecast and 6/18 analysis times for cumulative parameters
with 12h forecast. This makes querying the data through smartmet server
quite complex.

This commit introduces a view to the data that changes the times so that
it looks like we only have a single forecast which has analysis_time =
'2019-01-01T00:00:00' and maximum leadtime is 365 days.

This functionality relies on the fact that smartmet server does not check
the metadata from each grib but trusts whatever is defined in database.